Fix compilation warning on windows

+++ src/gui/src/ServerConfig.h
@@ -63,8 +63,8 @@ class ServerConfig : public BaseConfig
         bool ignoreAutoConfigClient() const { return m_IgnoreAutoConfigClient; }
         bool enableDragAndDrop() const { return m_EnableDragAndDrop; }
         bool clipboardSharing() const { return m_ClipboardSharing; }
-        size_t clipboardSharingSize() const { return m_ClipboardSharingSize; }
-        static size_t defaultClipboardSharingSize();
+        int clipboardSharingSize() const { return m_ClipboardSharingSize; }
+        static int defaultClipboardSharingSize();
 
         void saveSettings();
         void loadSettings();
@@ -94,7 +94,7 @@ class ServerConfig : public BaseConfig
         void setIgnoreAutoConfigClient(bool on) { m_IgnoreAutoConfigClient = on; }
         void setEnableDragAndDrop(bool on) { m_EnableDragAndDrop = on; }
         void setClipboardSharing(bool on) { m_ClipboardSharing = on; }
-        void setClipboardSharingSize(size_t size);
+        void setClipboardSharingSize(int size);
         QList<bool>& switchCorners() { return m_SwitchCorners; }
         std::vector<Hotkey>& hotkeys() { return m_Hotkeys; }
 
@@ -128,7 +128,7 @@ class ServerConfig : public BaseConfig
         bool m_IgnoreAutoConfigClient;
         bool m_EnableDragAndDrop;
         bool m_ClipboardSharing;
-        size_t m_ClipboardSharingSize;
+        int m_ClipboardSharingSize;
         MainWindow* m_pMainWindow;
 };
